Fixed c_intStr() prototype, updated sources list in Makefile to include StringConvert.c (hooray!).
I finished comparing the output from test_likepy and Robert's 1d Python test
and the results are the same.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@11507 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ test/studies/madness/dinan/mad_chapel/StringConvert.c
@@ -5,7 +5,7 @@ char * c_realStr(char * buf, char * fmt, double x) {
     return buf;
 }
 
-char * c_intStr(char * buf, char * fmt, double x) {
+char * c_intStr(char * buf, char * fmt, int x) {
     sprintf(buf, fmt, x);
     return buf;
 }
